Cleaned up and merged files
