+++ src/main/java/baritone/utils/schematic/format/defaults/LitematicaSchematic.java
@@ -18,25 +18,25 @@
 package baritone.utils.schematic.format.defaults;
 
 import baritone.utils.schematic.StaticSchematic;
-import baritone.utils.schematic.schematica.LitematicaBitArray;
 import net.minecraft.block.*;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.nbt.ListNBT;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.state.Property;
 import net.minecraft.util.registry.Registry;
+import org.apache.commons.lang3.Validate;
 
+import javax.annotation.Nullable;
 import java.util.*;
 
 /**
- * @author fencingparry4
+ * @author Emerson
  * @since 12/27/2020
  */
 public final class LitematicaSchematic extends StaticSchematic {
-    private String regionName;
 
     public LitematicaSchematic(CompoundNBT nbt) {
-        regionName = (String) nbt.getCompound("Regions").keySet().toArray()[0];
+        String regionName = (String) nbt.getCompound("Regions").keySet().toArray()[0];
         this.x = Math.abs(nbt.getCompound("Regions").getCompound(regionName).getCompound("Size").getInt("x"));
         this.y = Math.abs(nbt.getCompound("Regions").getCompound(regionName).getCompound("Size").getInt("y"));
         this.z = Math.abs(nbt.getCompound("Regions").getCompound(regionName).getCompound("Size").getInt("z"));
@@ -106,6 +106,104 @@ public final class LitematicaSchematic extends StaticSchematic {
             return state.with(property, parsed.get());
         } else {
             throw new IllegalArgumentException("Invalid value for property " + property);
+        }
+    }
+
+    /** LitematicaBitArray class from litematica */
+    private static class LitematicaBitArray
+    {
+        /** The long array that is used to store the data for this BitArray. */
+        private final long[] longArray;
+        /** Number of bits a single entry takes up */
+        private final int bitsPerEntry;
+        /**
+         * The maximum value for a single entry. This also works as a bitmask for a single entry.
+         * For instance, if bitsPerEntry were 5, this value would be 31 (ie, {@code 0b00011111}).
+         */
+        private final long maxEntryValue;
+        /** Number of entries in this array (<b>not</b> the length of the long array that internally backs this array) */
+        private final long arraySize;
+
+        public LitematicaBitArray(int bitsPerEntryIn, long arraySizeIn, @Nullable long[] longArrayIn)
+        {
+            Validate.inclusiveBetween(1L, 32L, (long) bitsPerEntryIn);
+            this.arraySize = arraySizeIn;
+            this.bitsPerEntry = bitsPerEntryIn;
+            this.maxEntryValue = (1L << bitsPerEntryIn) - 1L;
+
+            if (longArrayIn != null)
+            {
+                this.longArray = longArrayIn;
+            }
+            else
+            {
+                this.longArray = new long[(int) (roundUp((long) arraySizeIn * (long) bitsPerEntryIn, 64L) / 64L)];
+            }
+        }
+
+        public void setAt(long index, int value)
+        {
+            Validate.inclusiveBetween(0L, this.arraySize - 1L, (long) index);
+            Validate.inclusiveBetween(0L, this.maxEntryValue, (long) value);
+            long startOffset = index * (long) this.bitsPerEntry;
+            int startArrIndex = (int) (startOffset >> 6); // startOffset / 64
+            int endArrIndex = (int) (((index + 1L) * (long) this.bitsPerEntry - 1L) >> 6);
+            int startBitOffset = (int) (startOffset & 0x3F); // startOffset % 64
+            this.longArray[startArrIndex] = this.longArray[startArrIndex] & ~(this.maxEntryValue << startBitOffset) | ((long) value & this.maxEntryValue) << startBitOffset;
+
+            if (startArrIndex != endArrIndex)
+            {
+                int endOffset = 64 - startBitOffset;
+                int j1 = this.bitsPerEntry - endOffset;
+                this.longArray[endArrIndex] = this.longArray[endArrIndex] >>> j1 << j1 | ((long) value & this.maxEntryValue) >> endOffset;
+            }
+        }
+
+        public int getAt(long index)
+        {
+            Validate.inclusiveBetween(0L, this.arraySize - 1L, (long) index);
+            long startOffset = index * (long) this.bitsPerEntry;
+            int startArrIndex = (int) (startOffset >> 6); // startOffset / 64
+            int endArrIndex = (int) (((index + 1L) * (long) this.bitsPerEntry - 1L) >> 6);
+            int startBitOffset = (int) (startOffset & 0x3F); // startOffset % 64
+
+            if (startArrIndex == endArrIndex)
+            {
+                return (int) (this.longArray[startArrIndex] >>> startBitOffset & this.maxEntryValue);
+            }
+            else
+            {
+                int endOffset = 64 - startBitOffset;
+                return (int) ((this.longArray[startArrIndex] >>> startBitOffset | this.longArray[endArrIndex] << endOffset) & this.maxEntryValue);
+            }
+        }
+
+
+        public long size()
+        {
+            return this.arraySize;
+        }
+
+        public static long roundUp(long number, long interval)
+        {
+            if (interval == 0)
+            {
+                return 0;
+            }
+            else if (number == 0)
+            {
+                return interval;
+            }
+            else
+            {
+                if (number < 0)
+                {
+                    interval *= -1;
+                }
+
+                long i = number % interval;
+                return i == 0 ? number : number + interval - i;
+            }
         }
     }
 }
